feat(wallet): Top up & Dana Bos wallet flagging (#132)
* wip: wallet log 🔥

* wip 🔥

* feat: wallet log 🔥

* refactor (wallet): move controller to one folder

* wip 🔥

* feat: Top up wallet 🔥

* flag wallet as dana bos 💰

+++ app/Http/Controllers/Wallet/WalletController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Wallet;
 
+use App\Http\Controllers\Controller;
 use App\Models\Wallet;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
@@ -23,8 +24,9 @@ class WalletController extends Controller
      */
     public function create()
     {
-        $title = "Tambah Dompet";
-        return view('pages.wallet.create', compact('title'));
+        $data['title'] = "Tambah Dompet";
+        $data['danabos'] = Wallet::where('danabos', 1)->count();
+        return view('pages.wallet.create', $data);
     }
 
     /**
@@ -33,33 +35,26 @@ class WalletController extends Controller
     public function store(WalletRequest $request)
     {
         DB::beginTransaction();
+        $DanaBosCount = Wallet::where('danabos', 1)->count();
 
         try {
-
             $wallet                 = new Wallet();
             $wallet->school_id      = session('school_id');
             $wallet->name           = $request->name;
             $wallet->init_value     = formatAngka($request->init_value);
+            if ($request->has('danabos') and $DanaBosCount == 0) {
+                $wallet->danabos = true;
+            }
             $wallet->last_balance   = 0;
             $wallet->save();
 
             DB::commit();
-
-
         } catch (\Throwable $th) {
             DB::rollBack();
             return redirect()->route('wallet.index')->withToastError('Eror Simpan Dompet!');
         }
 
         return redirect()->route('wallet.index')->withToastSuccess('Berhasil Simpan Dompet!');
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(string $id)
-    {
-        //
     }
 
     /**
@@ -79,14 +74,12 @@ class WalletController extends Controller
         DB::beginTransaction();
 
         try {
-
             $wallet->school_id  = session('school_id');
             $wallet->name       = $request->name;
             $wallet->init_value = formatAngka($request->init_value);
             $wallet->save();
 
             DB::commit();
-
         } catch (\Throwable $th) {
             DB::rollBack();
             return redirect()->route('wallet.index')->withToastError('Eror Simpan Dompet!');
@@ -102,15 +95,19 @@ class WalletController extends Controller
     {
        DB::beginTransaction();
 
-       try {
+       if ($wallet->balance > 0) {
+            return response()->json([
+                'msg' => 'Tidak bisa menghapus Dompet!'
+            ]);
+       }
 
+       try {
         $wallet->delete();
         DB::commit();
 
         return response()->json([
             'msg' => 'Berhasil Hapus Dompet!'
         ], 200);
-
        } catch (\Throwable $th) {
             DB::rollBack();
             return response()->json([
